rename : redisUtil -> RedisService로 이름 수정
