+++ src/main/java/matchingGoal/matchingGoal/common/auth/JwtTokenProvider.java
@@ -5,7 +5,7 @@ import io.jsonwebtoken.security.Keys;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import matchingGoal.matchingGoal.common.type.ErrorCode;
-import matchingGoal.matchingGoal.common.util.RedisUtil;
+import matchingGoal.matchingGoal.common.service.RedisService;
 import matchingGoal.matchingGoal.member.exception.InvalidTokenException;
 import org.springframework.stereotype.Component;
 
@@ -19,7 +19,7 @@ import java.util.Map;
 @RequiredArgsConstructor
 public class JwtTokenProvider {
 
-    private final RedisUtil redisUtil;
+    private final RedisService redisService;
     private static final String BEARER_TYPE = "Bearer ";
     private static final String BLACK_TOKEN_PREFIX = "BLACK: ";
     private static final Long accessTokenExpirationTimeInSeconds = 30 * 60 * 1000L;
@@ -56,24 +56,24 @@ public class JwtTokenProvider {
     }
 
     public void saveRefreshToken(String email, String token){
-        redisUtil.setData(getRefreshTokenKey(email), token, refreshTokenExpirationTimeInSeconds);
+        redisService.setData(getRefreshTokenKey(email), token, refreshTokenExpirationTimeInSeconds);
     }
 
     public void deleteToken(String email) {
         String rtKey = getRefreshTokenKey(email);
-        if (redisUtil.getData(rtKey) == null)
+        if (redisService.getData(rtKey) == null)
             throw new InvalidTokenException(ErrorCode.INVALID_TOKEN);
 
-        redisUtil.deleteData(rtKey);
+        redisService.deleteData(rtKey);
     }
 
     public void setBlacklist(String token) {
-        redisUtil.setData(BLACK_TOKEN_PREFIX + getEmail(token), token, refreshTokenExpirationTimeInSeconds);
+        redisService.setData(BLACK_TOKEN_PREFIX + getEmail(token), token, refreshTokenExpirationTimeInSeconds);
     }
 
     public boolean validateToken(String token) {
         try {
-            if(redisUtil.getData(BLACK_TOKEN_PREFIX + getEmail(token)) != null )
+            if(redisService.getData(BLACK_TOKEN_PREFIX + getEmail(token)) != null )
                 return false;
 
             Jwts.parser().setSigningKey(secretKey).parseClaimsJws(getRawToken(token));
